Refactor UpstreamProject into its own class
